done with service and testing

+++ carservice/src/main/java/com/xloop/carservice/ICar.java
@@ -2,6 +2,6 @@ package com.xloop.carservice;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICar extends JpaRepository<Car, Integer> {
+public interface ICar extends JpaRepository<Car, Long> {
     
 }
